% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BikeWise-package.R
\docType{package}
\name{BikeWise-package}
\alias{BikeWise}
\alias{BikeWise-package}
\title{BikeWise: Bike Safe and Dry}
\description{
Plans a cycling route between two locations and checks the weather along the way to suggest the safest time to leave. Helps cyclists avoid rain by finding the next dry departure window within a 24-hour forecast.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sebkrbs/BikeWise}
  \item Report bugs at \url{https://github.com/sebkrbs/BikeWise/issues}
}

}
\author{
\strong{Maintainer}: Sebastian Krebs \email{sebasti.k24@gmail.com}

}
\keyword{internal}
