% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\docType{data}
\name{preset_titles}
\alias{preset_titles}
\title{Location preset labels}
\format{
An object of class \code{character} of length 8.
}
\usage{
preset_titles
}
\description{
A named character vector mapping the eight preset location keys to their
display names. Used as valid values for the \code{label} argument in
\code{\link{save_location}} and \code{\link{get_locations}}.
}
\keyword{datasets}
