% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_bikewise}
\alias{run_bikewise}
\title{Launch the BikeWise Shiny app}
\usage{
run_bikewise()
}
\value{
No return value. Starts the Shiny app in your browser.
}
\description{
Launch the BikeWise Shiny app
}
\examples{
\dontrun{
run_bikewise()
}

}
